Moved serial over to module, useful links on sharing.
