+++ firmware/src/serial.rs
@@ -1,0 +1,96 @@
+
+static mut USB_BUS: Option<UsbBusAllocator<UsbBusType>> = None;
+static mut USB_SERIAL: Option<usbd_serial::SerialPort<UsbBusType>> = None;
+static mut USB_DEVICE: Option<UsbDevice<UsbBusType>> = None;
+
+// On the whole sharing; 
+// https://github.com/rust-embedded/wg/issues/294#issuecomment-454425980
+// https://github.com/rust-embedded/wg/issues/294#issuecomment-456416517
+// https://github.com/rust-embedded/wg/issues/294#issuecomment-456742114
+// https://docs.rs/cortex-m/latest/cortex_m/interrupt/fn.free.html
+
+
+// SerialPort implements a stream.
+// CdcAcmClass implements packets...
+
+use stm32f1xx_hal::pac::{self, interrupt, Interrupt, NVIC};
+use stm32f1xx_hal::prelude::*;
+use stm32f1xx_hal::usb::{Peripheral, UsbBus, UsbBusType};
+use usb_device::{bus::UsbBusAllocator, prelude::*};
+use usbd_serial::{SerialPort, USB_CLASS_CDC};
+
+use core::cell::RefCell;
+use cortex_m::interrupt::Mutex;
+static SHARED: Mutex<RefCell<Option<usize>>> = Mutex::new(RefCell::new(None));
+
+pub struct Serial {
+}
+
+impl Serial {
+    pub fn new(usb: Peripheral) -> Self {
+
+        // Unsafe to allow access to static variables
+        unsafe {
+            let bus = UsbBus::new(usb);
+
+            USB_BUS = Some(bus);
+
+            USB_SERIAL = Some(SerialPort::new(USB_BUS.as_ref().unwrap()));
+
+            let usb_dev = UsbDeviceBuilder::new(USB_BUS.as_ref().unwrap(), UsbVidPid(0x16c0, 0x27dd))
+                .manufacturer("Fake company")
+                .product("Serial port")
+                .serial_number("TEST")
+                .device_class(USB_CLASS_CDC)
+                .build();
+
+            USB_DEVICE = Some(usb_dev);
+        }
+
+        unsafe {
+            NVIC::unmask(Interrupt::USB_HP_CAN_TX);
+            NVIC::unmask(Interrupt::USB_LP_CAN_RX0);
+        }
+        Serial{}
+    }
+}
+
+
+
+
+#[interrupt]
+fn USB_HP_CAN_TX() {
+    usb_interrupt();
+}
+
+#[interrupt]
+fn USB_LP_CAN_RX0() {
+    usb_interrupt();
+}
+
+fn usb_interrupt() {
+    let usb_dev = unsafe { USB_DEVICE.as_mut().unwrap() };
+    let serial = unsafe { USB_SERIAL.as_mut().unwrap() };
+
+    if !usb_dev.poll(&mut [serial]) {
+        return;
+    }
+
+    let mut buf = [0u8; 8];
+
+    match serial.read(&mut buf) {
+        Ok(count) if count > 0 => {
+            // Echo back in upper case
+            for c in buf[0..count].iter_mut() {
+                if 0x61 <= *c && *c <= 0x7a {
+                    *c &= !0x20;
+                }
+            }
+
+            serial.write(&buf[0..count]).ok();
+        }
+        _ => {}
+    }
+    /*
+    */
+}
